MV3.0, Extended Field Coding and ODO format support
Set Reader/Writer version R/W_T4T_Vno to 3.0
Extend all offsets, lengths, data pointers to 32-bit variables
to address memories larger than 32kB
Support Extended Field Coding by default if MLc>255 and/or MLe>256, even for MV2.0
For MV3.0 detected Tags, support ODO and DDO (BLE-TLV support) coding with
new ReadBinary and UpdateBinary INS code, Extended Data Structure
For addressing offsets below 0x8000, use MV2.0 commands for both MV2.0 and MV3.0 Tags
For addressing offsets above 0x8000 (i.e. MV3.0 Tag detected), use new MV3.0 INS code commands

+++ src/nfc/include/rw_api.h
@@ -965,7 +965,7 @@ extern tNFC_STATUS RW_T4tReadNDef(void);
 **                  NFC_STATUS_FAILED if T4T is busy or other error
 **
 *******************************************************************************/
-extern tNFC_STATUS RW_T4tUpdateNDef(uint16_t length, uint8_t* p_data);
+extern tNFC_STATUS RW_T4tUpdateNDef(uint32_t length, uint8_t* p_data);
 
 /*****************************************************************************
 **
